Fixed registration of default project resolver
I actually extended the resolvers class just to make all the resolvers sit in one place. Looks  nicer.
